+++ src/include/hosted/misc.hpp
@@ -29,6 +29,7 @@ extern std::string basename(const std::string& s);
 extern std::string basename(const char * const s);
 
 extern void hex_dump(std::ostream& out, const void *obj, const std::size_t size);
+extern std::string to_hex_string(const void *obj, const std::size_t size);
 
 /* ---------------------------------------------------------------------- */
 /*  */